chore(website): upgrade @rstack-dev/doc-ui to 1.13.3
