+++ website/package.json
@@ -16,7 +16,7 @@
     "@rspress/plugin-client-redirects": "^2.0.9",
     "@rspress/plugin-rss": "^2.0.9",
     "@rspress/plugin-sitemap": "^2.0.9",
-    "@rstack-dev/doc-ui": "^1.13.2",
+    "@rstack-dev/doc-ui": "1.13.3",
     "@shikijs/transformers": "^4.0.2",
     "@types/node": "^24.12.2",
     "@types/react": "^19.2.14",
